+++ src/components/articleInfo/articleInfo.js
@@ -93,6 +93,7 @@ export default function main(_) {
         })
     })();
 
+    /**
      * 设置文章标题-iconfont
      */
     (() => {