+++ src/components/articleInfo/articleInfo.js
@@ -60,11 +60,8 @@ export default function main(_) {
         _.__timeIds.articleInfoClassTId = window.setInterval(() => {
             let obj = $('#BlogPostCategory').find('a');
             if (obj.length > 0) {
-                $.each(obj, (i) => {
-                    let tag = $(obj[i]);
-                    tag.prepend('<span class="iconfont icon-marketing_fill"></span>');
-                    $('#articleInfo').append('<a href="'+tag.attr('href')+'" target="_blank"><span class="article-info-tag article-tag-class-color">'+(tag.text())+'</span></a>');
-                });
+                _.__tools.htmlReplace('#BlogPostCategory', /,/g, '')
+                _.__tools.articleInfo(obj, 1)
                 _.__tools.setDomHomePosition();
                 _.__tools.clearIntervalTimeId(_.__timeIds.articleInfoClassTId);
             }
@@ -78,15 +75,45 @@ export default function main(_) {
         _.__timeIds.articleInfoTagTId = window.setInterval(() => {
             let obj = $('#EntryTag').find('a');
             if (obj.length > 0) {
-                $.each(obj, (i) => {
-                    let tag = $(obj[i]);
-                    tag.prepend('<span class="iconfont icon-label_fill"></span>');
-                    $('#articleInfo').append('<a href="'+tag.attr('href')+'" target="_blank"><span class="article-info-tag article-tag-color">'+(tag.text())+'</span></a>');
-                });
+                _.__tools.htmlReplace('#EntryTag', /,/g, '')
+                _.__tools.articleInfo(obj, 2)
                 _.__tools.setDomHomePosition();
                 _.__tools.clearIntervalTimeId(_.__timeIds.articleInfoTagTId);
             }
         }, 1000);
     })();
+
+    /**
+     * 设置文章引用 | 扩展markdown语法
+     */
+    (() => {
+        $('.blogpost-body p').html((i, c) => {
+            if (/^\?&gt;/.test(c)) return '<p class="tip">' + c.slice(5).trim() + '</p>'
+            if (/^!&gt;/.test(c)) return '<p class="warn">' + c.slice(5).trim() + '</p>'
+        })
+    })();
+
+    /**
+     * 设置文章标题-iconfont
+     */
+    (() => {
+        let titleInfo = $('#cnblogs_post_body').find(':header')
+        if (_.__config.articleContent.prefixIcon.enable && titleInfo.length > 0) {
+            _.__tools.dynamicLoadingJs(_.__config.articleContent.prefixIcon.options.link).then(r => {
+                let iconfonts = _.__config.articleContent.prefixIcon.options.iconfontArr
+                titleInfo.html((i, c) => {
+                    let arr = []
+                    let num = Math.floor(Math.random() * (iconfonts.length - i) + i)
+                    if (arr.indexOf(num) == -1) {
+                        arr.push(num)
+                        $('<svg class="icon"> <use xlink:href="#icon-' + iconfonts[num] + '"></use></svg>').prependTo(titleInfo[i])
+                    } else {
+                        i--
+                    }
+                })
+            }).catch(e => console.error('iconfont.js', e))
+        }
+    })()
+
 
 }