+++ src/components/comment/comment.js
@@ -8,6 +8,23 @@
 import defaultAvatarImg from './../../images/webp/default_avatar.webp';
 
 export default function main(_) {
+
+    let avatarAnimate = () => {
+        // 头像动效
+        if (_.__config.animate.avatar.enable) {
+            let authorAvatar   = $('.author_avatar');
+            let feedbackAvatar = $('.feedbackAvatar img');
+
+            if (!authorAvatar.hasClass('img-rounded')) {
+                authorAvatar.addClass('img-rounded').css('border-radius', '50%');
+            }
+
+            if (!feedbackAvatar.hasClass('img-rounded')) {
+                feedbackAvatar.addClass('img-rounded').css('border-radius', '50%');
+            }
+        }
+    }
+
     let setComment = () => {
         let feedbackItem = $('.feedbackItem');
         if (feedbackItem.length > 0) {
@@ -40,17 +57,67 @@ export default function main(_) {
             });
             $(feedbackItem[0]).css('padding-top', '0');
             $(feedbackItem[feedbackItem.length - 1]).css('padding-bottom', '0');
+
+            avatarAnimate();
         }
     }
-    $(document).ajaxSuccess(function (event, xhr, settings) {
-        if (settings.url.includes("GetComments.aspx")) {
-            _.__tools.clearIntervalTimeId(_.__timeIds.commentTId);
-        }
-    });
+
+
+    let addComment = () => {
+        let userBlogAddress = $(".comment_my_posted a").attr('href'),
+            userName = $(".comment_my_posted a").text(),
+            commentInfo = $(".bq_post_comment").text();
+
+
+       let comment =  `<div class="feedbackItem" style="padding-bottom: 0px;">
+                        <div class="feedbackAvatar">
+                            <a href="${userBlogAddress}" target="_blank">
+                                <img src="${defaultAvatarImg}">
+                            </a>
+                        </div>
+                        <div class="feedbackListSubtitle ${ window.isBlogOwner && 'feedbackListSubtitle-louzhu'}">
+                            ${ window.isBlogOwner && `[<span class="louzhu">楼主</span>]`}
+                            <span class="comment_date">${(new Date).toLocaleString().replace(/\//g,'-')}</span>
+                            <a id="a_comment_author_5168811" href="${userBlogAddress}" target="_blank">${userName}</a>
+                        </div>
+                        <div class="feedbackCon">
+                            <div id="comment_body_5168811" data-format-type="Markdown" class="blog_comment_body cnblogs-markdown">
+                                <p>${commentInfo}</p>
+                            </div>
+                        </div>
+                    </div>`
+
+        $("#blog-comments-placeholder").append(comment)
+        $(".comment_my_posted").remove();
+
+        avatarAnimate();
+    }
+
     _.__timeIds.commentTId = window.setInterval(() =>{
         if ($('.feedbackItem').length > 0) {
             setComment();
             _.__tools.clearIntervalTimeId(_.__timeIds.commentTId);
         }
-    },1000);
+    }, 1000);
+
+    $(document).ajaxSuccess(function (event, xhr, settings) {
+        // 评论重新排序
+        if (settings.url.includes('GetComments.aspx')) {
+             _.__tools.clearIntervalTimeId(_.__timeIds.commentTId);
+            setComment()
+        }
+
+        // 新增评论
+        if (settings.url.includes('PostComment/Add.aspx')) addComment()
+
+        // 删除评论
+        if (settings.url.includes('comment/DeleteComment.aspx')) {
+            let commentId = JSON.parse(settings?.data)?.commentId;
+            $(`#comment_body_${commentId}`).parent().parent().remove()
+            $(".feedbackItem:last").css("padding-bottom", "0")
+        }
+
+        avatarAnimate();
+    });
+
 }