+++ src/components/sidebar/sidebar.js
@@ -34,7 +34,13 @@ export default function main(_) {
 
         // ------- 设置头像 -------
         let blogAvatar = _.__config.info.avatar ? _.__config.info.avatar : defaultAvatarImg;
-        $('#menuBlogAvatar').append("<img class='img-responsive img-rounded' alt='用户头像' src='"+blogAvatar+"'>");
+        $('#menuBlogAvatar').append("<img class='img-responsive' alt='用户头像' src='"+blogAvatar+"'>");
+
+        // 头像动效
+        if (_.__config.animate.avatar.enable) {
+            $('#menuBlogAvatar').css('border-radius', '50%')
+            $('#menuBlogAvatar .img-responsive').addClass('img-rounded')
+        }
 
         // ------- 设置侧边栏信息 -------
         $('.sidebar-title-msg').text(_.__config.sidebar.titleMsg);
@@ -338,5 +344,25 @@ export default function main(_) {
                 }
             }, 300);
         });
+    })();
+      
+    /**
+     * 设置是否默认展开菜单栏
+     */
+    (() => {
+        let submenuShow = (obj)  => {
+            obj.show().parent('li.sidebar-dropdown').addClass('active');
+        }
+        if (_.__config.sidebar.submenu.pointsRank) submenuShow($('#sb-sidebarScorerank'));
+        if (_.__config.sidebar.submenu.latestPosts) submenuShow( $('#sb-sidebarRecentposts'));
+        if (_.__config.sidebar.submenu.myTags) submenuShow($('#sb-toptags'));
+        if (_.__config.sidebar.submenu.postsClassify) submenuShow($('#sb-classify'));
+        if (_.__config.sidebar.submenu.articleClassify) submenuShow($('#sb-ArticleCategory'));
+        if (_.__config.sidebar.submenu.readRank) submenuShow($('#sb-topview'));
+        if (_.__config.sidebar.submenu.recommendRank) submenuShow($('#sb-topDiggPosts'));
+        if (_.__config.sidebar.submenu.postsArchive) submenuShow($('#sb-record'));
+        if (_.__config.sidebar.submenu.articleArchive) submenuShow($('#sb-articlearchive'));
+        if (_.__config.sidebar.submenu.latestComment) submenuShow($('#sb-recentComments'));
+        if (_.__config.sidebar.submenu.customList) submenuShow($('#customize-sidebar-menu .sidebar-submenu'));
     })();
 }