+++ docs/v2/Docs/changelog.md
@@ -6,6 +6,8 @@
 * 修复未开启找一找功能显示异常的问题 [(#374)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/374)
 * 文章分类和标签去除中间的逗号 [(#376)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/376)
 * 对titleMsg显示进行优化 [(#378)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/378)
+* 标题前缀图标 [(#379)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/379)
+* markdown extend [(#380)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/380)
 
 * 修复文章中文标题目录解析报错的问题 [(#377)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/377)
 * 修复解析自动识别代码语言的信息出错