+++ docs/v2/Docs/changelog.md
@@ -1,4 +1,18 @@
-## 2022.10.08 - v2.1.1
+## **** - v2.1.3
+* 合并播放器功能 [(#388)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/388)
+* 侧边栏配置隐藏 [(#387)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/387)
+* 评论优化 [(#384)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/384)
+* 头像旋转 [(#383)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/383)
+* 修复未开启找一找功能显示异常的问题 [(#374)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/374)
+* 文章分类和标签去除中间的逗号 [(#376)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/376)
+* 对titleMsg显示进行优化 [(#378)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/378)
+* 标题前缀图标 [(#379)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/379)
+* markdown extend [(#380)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/380)
+
+* 修复文章中文标题目录解析报错的问题 [(#377)](https://github.com/BNDong/Cnblogs-Theme-SimpleMemory/pull/377)
+* 修复解析自动识别代码语言的信息出错
+
+## 2022.10.08 - v2.1.2
 * 评估集成了一些网友提供的优化解决方案（感谢！）
 * 侧边栏增加网站统计
 * 由于cnzz开启了收费，废弃cnzz配置